Replaced `trace` with undefined `say`.
While using attendant in a plugin, I found that I wanted to trace the
progress of the attendant, but i didn't want to fuss with that silly
array, so I've defined the tracing function as an external function, to
be provided by the application developer.

The `trace` function has been renamed `say`.

For the plugin, the `say` function writes to a udp socket. With this
function linked into any of the participants, the plugin, the attendant,
it is easy to add a few say lines to see what is going on, without
having to add setup and teardown of logging files.

The same tracer can be used with the monitored out-of-process server as
well.