No restart if shutting down.
Now that we're chilling before a restart, it is very easy for a shutdown
to be requested right before a restart. Diving into the restart after
shutdown is requested is not working as I'd hoped. Sometimes it fails
before it gets to the reaper thread, so the shutdown switch is never
flipped.

I've added a shuttingdown flag to the process. If it is set when the
startup function wakes from a chill, it will skip the start and return
an failure. I'm also checking to see if shutdown is requested when I
go into the abend handler in the cases when starting fails before the
reaper is launched, that is the launcher reaps the failed child instead
of the reaper.

Closes #7.